Changed fibonacci program to conform to Betty style standards

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,22 +8,22 @@
 void fibonacci(void)
 {
 	int i;
-	unsigned long int a = 1;
-	unsigned long int b = 2;
+	unsigned long long a = 1;
+	unsigned long long b = 2;
 
-	printf("%lu, %lu, ", a, b);
+	printf("%llu, %llu, ", a, b);
 
 	for (i = 0; i < 96; i++)
 	{
 		if (a < b)
 		{
 			a = a + b;
-			printf("%lu", a);
+			printf("%llu", a);
 		}
 		else if (a > b)
 		{
 			b = b + a;
-			printf("%lu", b);
+			printf("%llu", b);
 		}
 
 		if (i != 95)
